#38 - Rename project, solutions and references to Baseline.FluentHttpExtensions
