+++ .github/workflows/release.yml
@@ -6,7 +6,7 @@ on:
             - "*"
 
 jobs:
-    build:
+    release:
         runs-on: ubuntu-latest
         steps:
             - uses: actions/checkout@v2
