+++ lib/Baseline.FluentHttpExtensions.Tests/Baseline.FluentHttpExtensions.Tests.csproj
@@ -2,10 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
-
     <IsPackable>false</IsPackable>
-
-    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
@@ -17,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FluentHttpExtensions.csproj" />
+    <ProjectReference Include="..\Baseline.FluentHttpExtensions.csproj" />
   </ItemGroup>
 
 </Project>
