+++ lib/Baseline.FluentHttpExtensions.Tests/Unit/HeaderExtensionsTests/WithBasicAuthTests.cs
@@ -3,7 +3,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace FluentHttpExtensions.Tests.Unit.HeaderExtensionsTests
+namespace Baseline.FluentHttpExtensions.Tests.Unit.HeaderExtensionsTests
 {
     public class WithBasicAuthTests : UnitTest
     {
