+++ lib/Baseline.FluentHttpExtensions.Tests/Unit/SendTriggeringExtensions/MakeRequestTests.cs
@@ -2,7 +2,7 @@ using System.Net;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
+namespace Baseline.FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
 {
     public class MakeRequestTests : UnitTest
     {
