+++ lib/Baseline.FluentHttpExtensions.Tests/Unit/SendTriggeringExtensions/ReadResponseAsStringTests.cs
@@ -1,15 +1,14 @@
 using System.Net.Http;
 using System.Threading.Tasks;
-using System.Xml.Serialization;
 using Moq;
 using Xunit;
 
-namespace FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
+namespace Baseline.FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
 {
-    public class ReadXmlResponseAsTests : UnitTest
+    public class ReadResponseAsStringTests : UnitTest
     {
         [Fact]
-        public async Task Checks_Success_Status_First()
+        public async Task Checks_Success_Response_First()
         {
             // Arrange.
             var mockMessageHandler = new Mock<HttpMessageHandler>();
@@ -17,33 +16,25 @@ namespace FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
             var request = new HttpRequest(RequestUrl, new HttpClient(mockMessageHandler.Object));
 
             // Act.
-            async Task Act() => await request.ReadXmlResponseAs<object>();
+            async Task Act() => await request.ReadResponseAsString();
 
             // Assert.
             await Assert.ThrowsAsync<HttpRequestException>(Act);
         }
 
         [Fact]
-        public async Task Returns_Xml_Successfully()
+        public async Task Returns_String_Content_Successfully()
         {
             // Arrange.
             var mockMessageHandler = new Mock<HttpMessageHandler>();
-            ConfigureMessageHandlerResultSuccess(mockMessageHandler, "<foo><bar>abc</bar></foo>", "application/xml");
+            ConfigureMessageHandlerResultSuccess(mockMessageHandler, "hello world", "text/plain");
             var request = new HttpRequest(RequestUrl, new HttpClient(mockMessageHandler.Object));
 
             // Act.
-            var response = await request.ReadXmlResponseAs<TestXmlObject>();
+            var response = await request.ReadResponseAsString();
 
             // Assert.
-            Assert.NotNull(response);
-            Assert.Equal("abc", response.Bar);
+            Assert.Equal("hello world", response);
         }
-    }
-
-    [XmlRoot("foo")]
-    public class TestXmlObject
-    {
-        [XmlElement("bar")]
-        public string Bar { get; set; }
     }
 }
