+++ lib/Baseline.FluentHttpExtensions.Tests/Unit/SendTriggeringExtensions/ReadXmlResponseAsTests.cs
@@ -1,14 +1,15 @@
 using System.Net.Http;
 using System.Threading.Tasks;
+using System.Xml.Serialization;
 using Moq;
 using Xunit;
 
-namespace FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
+namespace Baseline.FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
 {
-    public class ReadResponseAsStringTests : UnitTest
+    public class ReadXmlResponseAsTests : UnitTest
     {
         [Fact]
-        public async Task Checks_Success_Response_First()
+        public async Task Checks_Success_Status_First()
         {
             // Arrange.
             var mockMessageHandler = new Mock<HttpMessageHandler>();
@@ -16,25 +17,33 @@ namespace FluentHttpExtensions.Tests.Unit.SendTriggeringExtensions
             var request = new HttpRequest(RequestUrl, new HttpClient(mockMessageHandler.Object));
 
             // Act.
-            async Task Act() => await request.ReadResponseAsString();
+            async Task Act() => await request.ReadXmlResponseAs<object>();
 
             // Assert.
             await Assert.ThrowsAsync<HttpRequestException>(Act);
         }
 
         [Fact]
-        public async Task Returns_String_Content_Successfully()
+        public async Task Returns_Xml_Successfully()
         {
             // Arrange.
             var mockMessageHandler = new Mock<HttpMessageHandler>();
-            ConfigureMessageHandlerResultSuccess(mockMessageHandler, "hello world", "text/plain");
+            ConfigureMessageHandlerResultSuccess(mockMessageHandler, "<foo><bar>abc</bar></foo>", "application/xml");
             var request = new HttpRequest(RequestUrl, new HttpClient(mockMessageHandler.Object));
 
             // Act.
-            var response = await request.ReadResponseAsString();
+            var response = await request.ReadXmlResponseAs<TestXmlObject>();
 
             // Assert.
-            Assert.Equal("hello world", response);
+            Assert.NotNull(response);
+            Assert.Equal("abc", response.Bar);
         }
+    }
+
+    [XmlRoot("foo")]
+    public class TestXmlObject
+    {
+        [XmlElement("bar")]
+        public string Bar { get; set; }
     }
 }
